requirements.txt
asgiref==3.5.2
Django==4.0.3
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
haversine==2.7.0
django-postgres-timestamp-without-tz
python-dateutil